+++ client/lib/utils/algorithm_logic.js
