+++ client/lib/utils/cartographic_tools.js
