+++ client/lib/utils/latlong_conversion.js
