+++ package.json
@@ -44,6 +44,9 @@
     "eslint-plugin-prettier": "^3.0.0",
     "eslint-plugin-react": "^7.11.1",
     "prettier": "^1.15.3",
+    "react-redux": "^6.0.0",
+    "redux-logger": "^3.0.6",
+    "redux-saga": "^0.16.2",
     "webpack": "^4.0.0",
     "webpack-cli": "^3.1.2"
   },
